// src/app/orgunit-filter/models/orgunit.model.ts
export interface OrgUnitModel {
  selectionMode: string;
  selectedLevels: any[];
  showUpdateButton: boolean;
  selectedGroups: any[];
  orgUnits: any[];
  orgUnitLevels: any[];
  orgUnitGroups: any[];
  selectedOrgUnits: any[];
  selectedUserOrgUnit: any[];
  userOrgUnits: any[];
  type: string;
}